requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
async-timeout==3.0.1
attrs==22.2.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
discord.py==1.7.3
dnspython==2.2.1
et-xmlfile==1.1.0
ffmpeg==1.4
frozenlist==1.3.3
idna==3.4
mccabe==0.6.1
multidict==6.0.2
mutagen==1.46.0
mysql-connector==2.2.9
numpy==1.23.3
openpyxl==3.0.10
pandas==1.5.0
pycparser==2.20
pycryptodomex==3.17
pymongo==3.12.1
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.4
requests==2.28.1
six==1.15.0
toml==0.10.2
typing_extensions==4.1.1
urllib3==1.26.12
websockets==10.4
wrapt==1.12.1
xlrd==2.0.1
yarl==1.8.2
youtube-dl==2021.12.17
yt-dlp @ https://github.com/yt-dlp/yt-dlp/archive/master.tar.gz
